refactor(update-batch): moved ckeck_update to apps/views
